second_try_SOLOMA/requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
pydantic==2.5.0
python-dotenv==1.0.0
